Bring Schem into Alma
- Add Golem class.
It is instantiated by a content script injected into Alma tabs and can
in turn inject an iframe containing a repl based command line.
This can be toggled using the Go-Go-Golem command.

- Add debug logging options to the Schem interpreter.
These can also be turned on and off at runtime, via a Schem function.
e.g.: (setInterpreterOptions {"logArepInput" true})

- Add map-ish method to SchemMaps.
Makes reading or modifying of the values in a SchemMap easier.

- Add 'raw-loader' to webpack.
- Add module declarations for *.css and *.schem files.
 The content of *.css and *.schem files can now be imported as strings.

- Extract commandHistory into a seperate file.
- Move Key.enum to utils folder.
- Add comments, remove commented code.

TL;DR: This is not an atomic commit.  ._.